+++ spec/mpd_utils_spec.js
@@ -1,4 +1,5 @@
 /**
+ * @license
  * Copyright 2015 Google Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,8 +13,6 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
- * @fileoverview mpd_utils.js unit tests.
  */
 
 goog.require('shaka.dash.MpdUtils');
@@ -168,9 +167,105 @@ describe('MpdUtils', function() {
     });
   });
 
-  // TODO: Create more unit tests for this.
   describe('createTimeline', function() {
-    it('supports negative repetitions', function() {
+    it('works in normal case', function() {
+      var timepoints = [
+        createTimepoint(0, 10, 0),
+        createTimepoint(10, 10, 0),
+        createTimepoint(20, 10, 0)
+      ];
+      var result = [
+        { start: 0, end: 10 },
+        { start: 10, end: 20 },
+        { start: 20, end: 30 }
+      ];
+      checkTimepoints(timepoints, result, 1, 0);
+    });
+
+    it('handles null start time', function() {
+      var timepoints = [
+        createTimepoint(0, 10, 0),
+        createTimepoint(null, 10, 0),
+        createTimepoint(null, 10, 0)
+      ];
+      var result = [
+        { start: 0, end: 10 },
+        { start: 10, end: 20 },
+        { start: 20, end: 30 }
+      ];
+      checkTimepoints(timepoints, result, 1, 0);
+    });
+
+    it('handles gaps', function() {
+      var timepoints = [
+        createTimepoint(0, 10, 0),
+        createTimepoint(15, 10, 0)
+      ];
+      var result = [
+        { start: 0, end: 15 },
+        { start: 15, end: 25 }
+      ];
+      checkTimepoints(timepoints, result, 1, 0);
+    });
+
+    it('handles overlap', function() {
+      var timepoints = [
+        createTimepoint(0, 15, 0),
+        createTimepoint(10, 10, 0)
+      ];
+      var result = [
+        { start: 0, end: 10 },
+        { start: 10, end: 20 }
+      ];
+      checkTimepoints(timepoints, result, 1, 0);
+    });
+
+    it('handles repetitions', function() {
+      var timepoints = [
+        createTimepoint(0, 10, 5),
+        createTimepoint(60, 10, 0)
+      ];
+      var result = [
+        { start: 0, end: 10 },
+        { start: 10, end: 20 },
+        { start: 20, end: 30 },
+        { start: 30, end: 40 },
+        { start: 40, end: 50 },
+        { start: 50, end: 60 },
+        { start: 60, end: 70 }
+      ];
+      checkTimepoints(timepoints, result, 1, 0);
+    });
+
+    it('handles null repeat', function() {
+      var timepoints = [
+        createTimepoint(0, 10, 0),
+        createTimepoint(10, 10, null),
+        createTimepoint(20, 10, 0)
+      ];
+      var result = [
+        { start: 0, end: 10 },
+        { start: 10, end: 20 },
+        { start: 20, end: 30 }
+      ];
+      checkTimepoints(timepoints, result, 1, 0);
+    });
+
+    it('handles repetitions with gap', function() {
+      var timepoints = [
+        createTimepoint(0, 10, 2),
+        createTimepoint(35, 10, 0)
+      ];
+      var result = [
+        { start: 0, end: 10 },
+        { start: 10, end: 20 },
+        { start: 20, end: 35 },
+        { start: 35, end: 45 }
+      ];
+      checkTimepoints(timepoints, result, 1, 0);
+    });
+
+    it('handles negative repetitions', function() {
       var timepoints = [
         createTimepoint(0, 10, 0),
         createTimepoint(10, 10, -1),
@@ -186,7 +281,7 @@ describe('MpdUtils', function() {
       checkTimepoints(timepoints, result, 1, 0);
     });
 
-    it('supports negative repetitions with uneven border', function() {
+    it('handles negative repetitions with uneven border', function() {
       var timepoints = [
         createTimepoint(0, 10, 0),
         createTimepoint(10, 10, -1),
@@ -203,7 +298,7 @@ describe('MpdUtils', function() {
       checkTimepoints(timepoints, result, 1, 0);
     });
 
-    it('supports negative repetitions at end', function() {
+    it('handles negative repetitions at end', function() {
       var timepoints = [
         createTimepoint(0, 10, 0),
         createTimepoint(10, 5, -1)
@@ -217,12 +312,27 @@ describe('MpdUtils', function() {
       checkTimepoints(timepoints, result, 1, 25);
     });
 
+    it('ignores elements after null duration', function() {
+      var timepoints = [
+        createTimepoint(0, 10, 0),
+        createTimepoint(10, 10, 0),
+        createTimepoint(20, null, 0),
+        createTimepoint(30, 10, 0),
+        createTimepoint(40, 10, 0)
+      ];
+      var result = [
+        { start: 0, end: 10 },
+        { start: 10, end: 20 }
+      ];
+      checkTimepoints(timepoints, result, 1, 0);
+    });
+
     /**
      * Creates a new timepoint.
      *
-     * @param {number} start
-     * @param {number} dur
-     * @param {number} rep
+     * @param {?number} start
+     * @param {?number} dur
+     * @param {?number} rep
      */
     function createTimepoint(start, dur, rep) {
       var ret = new shaka.dash.mpd.SegmentTimePoint();