+++ spec/player_integration.js
@@ -1,5 +1,6 @@
 /**
- * Copyright 2014 Google Inc.
+ * @license
+ * Copyright 2015 Google Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,8 +13,6 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
- * @fileoverview Player integration tests.
  */
 
 goog.require('shaka.dash.MpdRequest');
@@ -75,15 +74,15 @@ describe('Player', function() {
 
     // Install polyfills.
     shaka.polyfill.installAll();
-
-    // Create a video tag.  This will be visible so that long tests do not
-    // create the illusion of the test-runner being hung.
-    video = createVideo();
-    document.body.appendChild(video);
   });
 
   beforeEach(function() {
+    // Create a video tag.  This will be visible so that long tests do not
+    // create the illusion of the test-runner being hung.
+    video = createVideo();
     video.autoplay = false;
+    document.body.appendChild(video);
+
     player = createPlayer(video);
     eventManager = new shaka.util.EventManager();
   });
@@ -96,12 +95,12 @@ describe('Player', function() {
       player = null;
       done();
     });
+
+    // Remove the video tag from the DOM.
+    document.body.removeChild(video);
   });
 
   afterAll(function() {
-    // Remove the video tag from the DOM.
-    document.body.removeChild(video);
-
     // Restore the timeout.
     jasmine.DEFAULT_TIMEOUT_INTERVAL = originalTimeout;
 
@@ -173,18 +172,6 @@ describe('Player', function() {
   });
 
   describe('selectVideoTrack', function() {
-    beforeEach(function(done) {
-      // On some browsers, using the same video tag for these test cause the
-      // tests to be flakely. So use a fresh video tag for each test.
-      player.destroy().then(function() {
-        document.body.removeChild(video);
-        video = createVideo();
-        document.body.appendChild(video);
-        player = createPlayer(video);
-        done();
-      });
-    });
-
     it('can set resolution before beginning playback', function(done) {
       player.load(newSource(plainManifest)).then(function() {
         var track = getVideoTrackByHeight(720);
@@ -305,14 +292,13 @@ describe('Player', function() {
 
       // Create a temporary shim to intercept and modify manifest info.
       var originalLoad = shaka.player.StreamVideoSource.prototype.load;
-      shaka.player.StreamVideoSource.prototype.load = function(
-          preferredLanguage) {
+      shaka.player.StreamVideoSource.prototype.load = function() {
         var sets = this.manifestInfo.periodInfos[0].streamSetInfos;
         var audioSet = sets[0].contentType == 'audio' ? sets[0] : sets[1];
         expect(audioSet.contentType).toBe('audio');
         // Remove the video set to speed things up.
         this.manifestInfo.periodInfos[0].streamSetInfos = [audioSet];
-        return originalLoad.call(this, preferredLanguage);
+        return originalLoad.call(this);
       };
 
       var audioStreamBuffer;
@@ -335,7 +321,7 @@ describe('Player', function() {
       }).then(function() {
         // Power through and consume the audio data quickly.
         player.setPlaybackRate(4);
-        return delay(5);
+        return waitForTargetTime(video, eventManager, 30, 30);
       }).then(function() {
         // Ensure that the browser has evicted the beginning of the stream.
         // Otherwise, this test hasn't reproduced the circumstances correctly.
@@ -435,6 +421,7 @@ describe('Player', function() {
 
     function streamStartupTest(playbackStartTime, seekTarget, done) {
       var source = newSource(plainManifest);
+      video.autoplay = true;
 
       player.setPlaybackStartTime(playbackStartTime);
 
@@ -443,25 +430,13 @@ describe('Player', function() {
       var originalLoad = shaka.player.StreamVideoSource.prototype.load;
       shaka.player.StreamVideoSource.prototype.load = function() {
         expect(this.manifestInfo).not.toBe(null);
-        this.manifestInfo.minBufferTime = 15;
+        this.manifestInfo.minBufferTime = 80;
         return originalLoad.call(this);
       };
 
-      var pollTimer = null;
-
       player.load(source).then(function() {
-        video.play();
-
         // Continue once we've buffered at least one segment.
-        var p = shaka.util.PublicPromise();
-        var pollBuffer = function() {
-          if (video && video.buffered.length == 1) {
-            window.clearInterval(pollTimer);
-            p.resolve();
-          }
-        };
-        pollTimer = window.setInterval(pollBuffer, 25);
-        return p;
+        return waitFor(30, function() { return video.buffered.length == 1; });
       }).then(function() {
         // Ensure we have buffered at least one segment but have not yet
         // started playback.
@@ -469,19 +444,21 @@ describe('Player', function() {
         expect(video.playbackRate).toBe(0);
         // Now seek back near the beginning.
         video.currentTime = seekTarget;
-        return delay(3);
+
+        // Once it seeks to the beginning, it will buffer the video and then
+        // start playing once it reaches its buffer goal.
+        return waitFor(30, function() { return video.playbackRate != 0; });
+      }).then(function() {
+        // Ensure that the video is actually moving.
+        return delay(1);
       }).then(function() {
         expect(video.buffered.length).toBe(1);
         expect(video.playbackRate).toBe(1);
-        expect(video.currentTime > seekTarget);
+        expect(video.currentTime).toBeGreaterThan(seekTarget);
         shaka.player.StreamVideoSource.prototype.load = originalLoad;
         done();
       }).catch(function(error) {
         shaka.player.StreamVideoSource.prototype.load = originalLoad;
-
-        if (pollTimer != null) {
-          window.clearTimeout(pollTimer);
-        }
 
         fail(error);
         done();
@@ -778,6 +755,102 @@ describe('Player', function() {
         done();
       });
     });
+
+    it('plays in reverse past the buffered area', function(done) {
+      var timestamp;
+      // Start in the second segment.
+      player.setPlaybackStartTime(10);
+      player.load(newSource(plainManifest)).then(function() {
+        timestamp = video.currentTime;
+        video.play();
+        return waitForMovement(video, eventManager);
+      }).then(function() {
+        expect(video.buffered.length).toBe(1);
+        expect(video.buffered.start(0)).toBeGreaterThan(5);
+        player.setPlaybackRate(-3.0);
+        return waitForMovement(video, eventManager);
+      }).then(function() {
+        return delay(4.0);
+      }).then(function() {
+        expect(video.currentTime).toBeLessThan(0.1);
+        done();
+      }).catch(function(error) {
+        fail(error);
+        done();
+      });
+    });
+
+    it('pauses while rewinding', function(done) {
+      var timestamp;
+      player.setPlaybackStartTime(45);
+      player.load(newSource(plainManifest)).then(function() {
+        video.play();
+        return waitForTargetTime(video, eventManager, 49, 6);
+      }).then(function() {
+        player.setPlaybackRate(-1.0);
+        return delay(2.0);
+      }).then(function() {
+        video.pause();
+        timestamp = video.currentTime;
+        return delay(3.0);
+      }).then(function() {
+        expect(video.paused).toBe(true);
+        expect(video.currentTime).toBe(timestamp);
+        done();
+      }).catch(function(error) {
+        fail(error);
+        done();
+      });
+    });
+
+    it('does not rewind while paused', function(done) {
+      var timestamp;
+      player.setPlaybackStartTime(45);
+      player.load(newSource(plainManifest)).then(function() {
+        video.play();
+        return waitForTargetTime(video, eventManager, 49, 6);
+      }).then(function() {
+        video.pause();
+        timestamp = video.currentTime;
+        return delay(3.0);
+      }).then(function() {
+        player.setPlaybackRate(-1.0);
+        return delay(2.0);
+      }).then(function() {
+        expect(video.paused).toBe(true);
+        expect(video.currentTime).toBe(timestamp);
+        done();
+      }).catch(function(error) {
+        fail(error);
+        done();
+      });
+    });
+
+    it('rewinds after pausing', function(done) {
+      var timestamp;
+      player.setPlaybackStartTime(45);
+      player.load(newSource(plainManifest)).then(function() {
+        video.play();
+        return waitForMovement(video, eventManager);
+      }).then(function() {
+        player.setPlaybackRate(-1.0);
+        return delay(2.0);
+      }).then(function() {
+        video.pause();
+        timestamp = video.currentTime;
+        return delay(3.0);
+      }).then(function() {
+        video.play();
+        return delay(2.0);
+      }).then(function() {
+        expect(video.paused).toBe(false);
+        expect(video.currentTime).toBeLessThan(timestamp);
+        done();
+      }).catch(function(error) {
+        fail(error);
+        done();
+      });
+    });
   });
 
   describe('getStats', function() {
@@ -863,6 +936,22 @@ describe('Player', function() {
         done();
       });
     });
+
+    it('respects the \'main\' attribute', function(done) {
+      // There are no Thai audio AdaptationSets in the MPD, but the English
+      // audio AdaptationSet is marked as main, so even though it is not the
+      // first AdaptationSet, it should be preferred.
+      player.configure({'preferredLanguage': 'th'});
+
+      player.load(newSource(languagesManifest)).then(function() {
+        var activeAudioTrack = getActiveAudioTrack();
+        expect(activeAudioTrack.lang).toBe('en');
+        done();
+      }).catch(function(error) {
+        fail(error);
+        done();
+      });
+    });
   });
 
   describe('configuring restrictions', function() {
@@ -935,6 +1024,45 @@ describe('Player', function() {
         done();
       });
     });
+
+    it('fires error if all tracks restricted while playing', function(done) {
+      player.removeEventListener('error', convertErrorToTestFailure, false);
+      var onError = jasmine.createSpy('onError');
+      player.addEventListener('error', onError, false);
+
+      player.load(newSource(plainManifest)).then(function() {
+        waitForMovement(video, eventManager);
+      }).then(function() {
+        var restrictions = player.getConfiguration()['restrictions'];
+        restrictions.maxBandwidth = 10000;
+        player.configure({'restrictions': restrictions});
+
+        expect(onError.calls.any()).toBe(true);
+        expect(player.getVideoTracks().length).toBe(0);
+
+        done();
+      }).catch(function(error) {
+        fail(error);
+        done();
+      });
+    });
+
+    it('fires error if all tracks restricted before playing', function(done) {
+      player.removeEventListener('error', convertErrorToTestFailure, false);
+      var onError = jasmine.createSpy('onError');
+      player.addEventListener('error', onError, false);
+
+      var restrictions = player.getConfiguration()['restrictions'];
+      restrictions.maxBandwidth = 10000;
+      player.configure({'restrictions': restrictions});
+
+      player.load(newSource(encryptedManifest)).then(function() {
+        fail();
+      }).catch(function(error) {
+        expect(onError.calls.any()).toBe(true);
+        done();
+      });
+    });
   });
 
   describe('interpretContentProtection', function() {
@@ -963,8 +1091,6 @@ describe('Player', function() {
       player.load(newSourceWithIcp(icp)).then(function() {
         video.play();
         return waitForMovement(video, eventManager);
-      }).then(function() {
-        return delay(0.5);
       }).then(function() {
         expect(video.currentTime).toBeGreaterThan(0.0);
         done();
@@ -996,7 +1122,7 @@ describe('Player', function() {
 
       player.load(newSourceWithIcp(icp)).then(function() {
         video.play();
-        return delay(0.5);
+        return waitForMovement(video, eventManager);
       }).then(function() {
         expect(licensePostProcessor.spy).toHaveBeenCalled();
         done();
@@ -1077,13 +1203,15 @@ describe('Player', function() {
 
       player.load(newSourceWithIcp(icp)).then(function() {
         video.play();
-        return delay(0.5);
+        return waitForMovement(video, eventManager);
       }).then(function() {
         expect(licensePreProcessor.spy).toHaveBeenCalled();
         // done() is called from the LicenseRequest.send() spy above.
       }).catch(function(error) {
-        fail(error);
-        done();
+        if (error.type != 'destroy') {
+          fail(error);
+          done();
+        }
       });
     });
   });
@@ -1384,7 +1512,12 @@ describe('Player', function() {
         return waitForMovement(video, eventManager);
       }).then(function() {
         video.currentTime = video.duration - 2;
-        return delay(6);
+        return waitFor(30, function() {
+          return video.currentTime > 0 && video.currentTime < 5;
+        }, function(error) {
+          error.message = 'Timeout waiting for loop, currentTime = ' +
+                          video.currentTime;
+        });
       }).then(function() {
         expect(video.currentTime).toBeGreaterThan(0);
         expect(video.currentTime).toBeLessThan(5);
@@ -1405,7 +1538,9 @@ describe('Player', function() {
             plainManifestDuration - 1 - 0.1);
         expect(video.currentTime).toBeLessThan(
             plainManifestDuration - 1 + 0.1);
-        return delay(5);
+        return waitFor(30, function() {
+          return video.currentTime > 0 && video.currentTime < 5;
+        });
       }).then(function() {
         expect(video.currentTime).toBeGreaterThan(0);
         expect(video.currentTime).toBeLessThan(5);
@@ -1422,7 +1557,9 @@ describe('Player', function() {
       player.load(newSource(plainManifest)).then(function() {
         // Note: the browser does not immediately set the video's duration.
         video.currentTime = plainManifestDuration;
-        return delay(5);
+        return waitFor(30, function() {
+          return video.currentTime > 0 && video.currentTime < 5;
+        });
       }).then(function() {
         expect(video.currentTime).toBeGreaterThan(0);
         expect(video.currentTime).toBeLessThan(5);
@@ -1440,7 +1577,9 @@ describe('Player', function() {
         return waitForMovement(video, eventManager);
       }).then(function() {
         video.currentTime = video.duration + 10;
-        return delay(4);
+        return waitFor(30, function() {
+          return video.currentTime > 0 && video.currentTime < 5;
+        });
       }).then(function() {
         expect(video.currentTime).toBeGreaterThan(0);
         expect(video.currentTime).toBeLessThan(5);
@@ -1450,6 +1589,143 @@ describe('Player', function() {
         done();
       });
     });
+  });
+
+  describe('buffering', function() {
+    // Tests that using a streamBufferSize < minBufferTime does not cause
+    // the player to hang after re-buffering. Issue #166.
+    it('does not stop playback after re-buffering', function(done) {
+      // Don't fail from fake errors.
+      player.removeEventListener('error', convertErrorToTestFailure, false);
+      player.addEventListener(
+          'error',
+          function(event) {
+            if (event.detail.message != 'Fake network failure.') {
+              fail(event.detail);
+            }
+          },
+          false);
+
+      // Set the initial streamBufferSize to be less than the default (so we
+      // can run the test faster) but larger than minBufferTime (so we don't
+      // potentially trigger a separate bug, see the test below).
+      player.configure({'streamBufferSize': 8});
+
+      var originalLoad = shaka.player.StreamVideoSource.prototype.load;
+      shaka.player.StreamVideoSource.prototype.load = function() {
+        expect(this.manifestInfo).not.toBe(null);
+        this.manifestInfo.minBufferTime = 2;
+        return originalLoad.call(this);
+      };
+
+      var originalSend = shaka.util.AjaxRequest.prototype.send;
+
+      player.load(newSource(plainManifest)).then(function() {
+        shaka.player.StreamVideoSource.prototype.load = originalLoad;
+        video.play();
+        return waitForMovement(video, eventManager);
+      }).then(function() {
+        player.configure({'streamBufferSize': 2});
+
+        // Start blocking network requests.
+        shaka.util.AjaxRequest.prototype.send = function() {
+          shaka.asserts.assert(this.xhr_ == null);
+          this.xhr_ = new XMLHttpRequest();
+          var error = new Error('Fake network failure.');
+          error.type = 'net';
+          error.status = 0;
+          error.url = this.url;
+          error.method = this.parameters.method;
+          error.body = this.parameters.body;
+          error.xhr = this.xhr_;
+          this.promise_.reject(error);
+          return this.promise_;
+        };
+        return delay(12);
+      }).then(function() {
+        // Ensure the player is in a buffering state.
+        expect(video.paused).toBe(true);
+        // Start allowing network requests again.
+        shaka.util.AjaxRequest.prototype.send = originalSend;
+        return waitForMovement(video, eventManager);
+      }).then(function() {
+        expect(video.paused).toBe(false);
+        done();
+      }).catch(function(error) {
+        shaka.player.StreamVideoSource.prototype.load = originalLoad;
+        shaka.util.AjaxRequest.prototype.send = originalSend;
+
+        fail(error);
+        done();
+      });
+    });
+
+    // Tests that using a streamBufferSize < minBufferTime does not cause
+    // the player to hang during startup. Issue #166.
+    it('does not stop playback during startup', function(done) {
+      player.configure({'streamBufferSize': 5});
+
+      var originalLoad = shaka.player.StreamVideoSource.prototype.load;
+      shaka.player.StreamVideoSource.prototype.load = function() {
+        expect(this.manifestInfo).not.toBe(null);
+        this.manifestInfo.minBufferTime = 7;
+        return originalLoad.call(this);
+      };
+
+      player.load(newSource(plainManifest)).then(function() {
+        video.play();
+        return waitForMovement(video, eventManager);
+      }).then(function() {
+        done();
+      }).catch(function(error) {
+        shaka.player.StreamVideoSource.prototype.load = originalLoad;
+
+        fail(error);
+        done();
+      });
+    });
+  });
+
+  describe('destroy', function() {
+    it('does not cause exceptions after load', function(done) {
+      // Satisfy Jasmine, which complains if tests have no expectations.
+      expect(true).toBe(true);
+      // Load the manifest, then wait.
+      player.load(newSource(encryptedManifest)).then(function() {
+        player.destroy();
+        // Wait 1 second for async exceptions.
+        // Jasmine will convert them to test failures.
+        delay(1).then(done);
+      });
+    });
+
+    it('does not cause exceptions immediately', function(done) {
+      // Satisfy Jasmine, which complains if tests have no expectations.
+      expect(true).toBe(true);
+      // Chrome complains if you don't catch all failed Promises.
+      player.load(newSource(encryptedManifest)).catch(function() {});
+      // Now destroy the player without letting it finish loading.
+      player.destroy();
+      // Wait 1 second for async exceptions.
+      // Jasmine will convert them to test failures.
+      delay(1).then(done);
+    });
+
+    for (var ms = 1; ms <= 1000; ms *= 5) {
+      it('does not cause exceptions after ' + ms + ' ms', function(ms, done) {
+        // Satisfy Jasmine, which complains if tests have no expectations.
+        expect(true).toBe(true);
+        // Chrome complains if you don't catch all failed Promises.
+        player.load(newSource(encryptedManifest)).catch(function() {});
+        // Now destroy the player after the specified delay.
+        delay(ms / 1000).then(function() {
+          player.destroy();
+          // Wait 1 additional second for async exceptions.
+          // Jasmine will convert them to test failures.
+          delay(1).then(done);
+        });
+      }.bind(this, ms));
+    }
   });
 
   // TODO(story 1970528): add tests which exercise PSSH parsing,